+++ Server/WoofWoofMaps.Bll/Models/PointModel.cs
@@ -1,3 +1,3 @@
 ﻿namespace WoofWoofMaps.Bll.Models;
 
-public record PointModel(decimal Latitude, decimal Longitude, DateTime Timestamp);
+public record PointModel(double Latitude, double Longitude, DateTime Timestamp);
